I changed the redundant logic in parseNewElement and gave clearer names
Since many local variables were called growingLists that could have lead to the confusion about whether any method is making changes to the list that it takes in as a parameter. Now, I my local variables are changed from growingLists to oldGrowingLists and newGrowingLists so that it is clear that methods aren't making changes to the lists that they take in as a parameter. When it comes to the parseNewElement method, the first check of whether the current index is out of bounds is redundant if I later have if(NEXT_INDEX < array.length) check.