Adding virtual methods to Animal.h file
At first I thought there was no reason for virtual methods to be in the .h file, but I have realized that when child classes don't want to change virtual methods of their parent classes, they should be able to simply add the function from the parent class into their vTableGlobal object instead of copy pasting the body of the function. In order for them to have access to these methods at compile time, virtual methods of the parent class need to be declared in the .h file.